Added hungry animal behaviour, made Camp keep track of the approximate center of the camp. Closes issue 16.

+++ Goblin Camp/src/Menu.cpp
@@ -514,7 +514,7 @@ MenuResult StockManagerMenu::Update(int x, int y) {
 		} else if (x == topX+48 && y == topY+1) {
 			if (scroll > 0) --scroll;
 		} else if (x == topX+48 && y == topY+48) {
-			if (scroll < (StockManager::Inst()->Producables()->size() / 3)-1) ++scroll;
+			if (scroll < (signed int)(StockManager::Inst()->Producables()->size() / 3)-1) ++scroll;
 		}
 		return MENUHIT;
 	}
